Fixes #303: fix several problems on load/save maps
There were a bunch of problems when saving and
loading projects with mapsheets.
ProjectViews were cloned for its use in MapSheets,
but they had a void name, which lead to problems
when saving and reloading the project. Now they
retain the same name.
Also the PV is now cloned when loading a saved
MapSheets layout.
A MapSheetFrameViewFactory has now been created
and registered so the MapSheetFrameView cand now
be restored like any other Frame when loading
the project. The MapSheetFrameView has also
been modified so it's saved in the gvp.
The XML entity from MapSheetsLayoutTemplate now
includes a field with the main frame grid name.
ScaleBars are also properly linked with their view
when restored.