ExtGex&ExtPM: forms xml file properly loaded.
When moving some data to extAudasaCommons, some
errors started to raise. I thinks this was due
to the fact that loading the library through
that project may lead to not finding the xml
files (they are looked for through the wrong
base dir). Now the InputStreams are created
in each extension and then passed to the
FormPanel constructor.