ExtPM&ExtGIA&ExtAC:reports images paths fixed.
When moving the libraries to AudasaCommons,
another problem raised with the reports,
because the path to the embedded images could
not be accessed the same way. This has been
fixed by changing some methods for accepting
an additional parameter with the extension
path, as well as adding a new variable in the
Jasper report.