extGIA: refs #980. Fixes NullPointerException
When there are no jobs and/or there are no "medicion_elemento" a
NullPointerException is raised.

Also IController.createValues does not handle correctly null values
so it should be converted to empty strings previously.

Not sure if createValues must handle it or no, so i change it in client
code to avoid touching the library
createValues