extELLE, siga, create-gvsig-portable. Removes 'legends' from ELLE
extELLE on SIGA should only know about maps, and not about legends. So
all the legends extensions (save, load, delete) are dropped. And the
maps extensiones (save, load, delete) are modified to handle also the
legends, as if always exists a group of leyends with the same name of
map