extNavTableForms: Multiple fk keys for vectorial embed tables
A vectorial table will be used as an embed table for a couple of different
parent layer. The embed table will be referenced with a composed fk key.
So a couple of changes are needed. In this change we encapsulate (making)
private two fields to avoid and the childs access them via accessor methods.
So if childs now that there is only one field in the pk nothing changes.

This will be only implemented for VectorialTableHandler